Revert breaking change

+++ FFXIVClientStructs/FFXIV/Client/UI/Agent/AgentTelepotTown.cs
@@ -23,7 +23,7 @@ public unsafe partial struct AgentTelepotTownData {
     [FieldOffset(0x0C), FixedSizeArray] internal FixedSizeArray64<AetheryteEntry> _entries;
 
     [FieldOffset(0x70E)] public byte SelectedAetheryte; // index selected in the aethernet list
-    [FieldOffset(0x710)] public uint Flags;
+    [FieldOffset(0x710)] public byte Flags; // TODO: uint
 }
 
 [StructLayout(LayoutKind.Explicit, Size = 0x18)]
